requirements.txt
crewai
crewai-tools
openai 
requests 
email-validator 
python-dotenv
apscheduler
